+++ flutter_sheet_localization_generator/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 2.0.1
+
+- Added multiline support
+
+## 2.0.0
+
+- Added version parameter to force generation
+- Added support for typed templated values
+- Added support for formatting rules for templated values
+
 ## 1.2.0
 
 - Upgraded analayzer dependency